test: 별도 데이터베이스 -> 기존 데이터베이스 사용하도록 수정
