+++ app/database.py
@@ -1,6 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, declarative_base
 import os
+from sqlalchemy import MetaData
 
 engine = create_engine(os.getenv("DATABASE_URL"))
 
@@ -15,3 +16,12 @@ def get_db():
         yield db
     finally:
         db.close()
+
+
+def clear_db():
+    meta = MetaData()
+    meta.reflect(bind=engine)
+    with engine.connect() as conn:
+        for table in reversed(meta.sorted_tables):
+            conn.execute(table.delete())
+        conn.commit()
