+++ app/tests/test_00_base.py
@@ -1,0 +1,9 @@
+import sys, os
+
+sys.path.append("app/")
+os.environ["DATABASE_URL"] = "sqlite:///:memory:"
+
+from database import Base, engine
+from models import *
+
+Base.metadata.create_all(bind=engine)
